+++ src/UI/Components/index.ts
@@ -1,3 +1,4 @@
 export * from "./Toggle";
 export * from "./TreeTable";
 export * from "./ResourceTable";
+export * from "./StackedProgressBar";