+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 - Disable edit and delete instance actions when not allowed by the lifecycle (#184)
 - Add resources tab to service inventory (#164)
 - Add inmanta favicon (#78)
+- Show resource deployment progress in the inventory view (#154)
 
 # Release 1.4.0 (2020-12-17)